Login-Android-RealDevice/requirements.txt
Appium_Python_Client==2.1.2
pytest==6.2.5
selenium==4.1.0
